separated moviesSaga - created a separate Saga and Slice file for downloading movie details
