+++ src/features/movies/moviesSaga.js
@@ -1,14 +1,10 @@
-import { put, call, select, takeEvery, delay, all, fork } from "redux-saga/effects";
+import { put, call, select, takeEvery, delay, fork } from "redux-saga/effects";
 import { fetchApiData } from "../../fetchApiData";
 import {
   showMovies,
-  getMovieDetails,
   fetchMoviesSuccess,
-  fetchMovieDetailsSuccess,
   fetchMoviesError,
-  fetchMovieDetailsError,
   setGenres,
-  fetchMovieDetailsLoading,
   fetchMoviesLoading,
 } from "./moviesSlice";
 import { selectContentType, totalPages, selectPage, contentType } from "../../common/Pagination/paginationSlice";
@@ -43,30 +39,7 @@ export function* fetchGenresHandler() {
   }
 }
 
-export function* fetchMovieDetailsHandler({ payload: id }) {
-  try {
-    yield put(fetchMovieDetailsLoading());
-    yield delay(2000);
-    const detailsApiData = `https://api.themoviedb.org/3/movie/${id}?api_key=6007bf485fd1645cfc7ab81654ba3228&language=en-US`;
-    const creditsApiData = `https://api.themoviedb.org/3/movie/${id}/credits?api_key=6007bf485fd1645cfc7ab81654ba3228&language=en-US`;
-
-    const [movieDetails, credits] = yield all([
-      call(fetchApiData, detailsApiData),
-      call(fetchApiData, creditsApiData),
-    ]);
-
-    yield put(fetchMovieDetailsSuccess({
-      ...movieDetails,
-      cast: credits.cast,
-      crew: credits.crew,
-    }));
-  } catch (error) {
-    yield put(fetchMovieDetailsError());
-  }
-}
-
 export function* moviesSaga() {
   yield fork(fetchMoviesHandler);
   yield takeEvery(showMovies.type, fetchMoviesHandler);
-  yield takeEvery(getMovieDetails.type, fetchMovieDetailsHandler);
 }
